requirements.txt
langgraph
langchain
langchain-anthropic
anthropic
mcp
pydantic
fastapi
uvicorn
langchain-openai
openai
openai-agents
python-dotenv